Modified the sphinx settings
It seems the manpages ``pelican-themes.en.1`` and
``pelican-themes.fr.1`` were only used for the Debian packages.

Unforunately, ``dh_manpages`` seems to support installation of manpages
depending of their languages, so i've changed the settings and only
``pelican-themes.1`` is generated, this avoids having to rename the
manpages before packaging...

The documentation about making themes for Pelican is now generated in
the file ``pelican-theming.1``.